Create single factory (#424)
